EOD Refactor & New Structure
EOD for marines/ODST has been updated to inherit from the marines helm, and a new structure has been created for how to handle special cert required helmets for non spartans.
