+++ src/Addons/34thPRC_ArmourStandard/data/marines2/config_weapons.hpp
@@ -30,9 +30,13 @@ class MA_Marine_BDU_NoShoulder: U_B_CombatUniform_mcam //MA_Armor
 {
 	class ItemInfo;
 };
+class ItemInfo;
+class MA_Helmet_Base; //MA_Armor Helmets.hpp
+class MA_EOD_CBRN_Attachment_NVG; //MA_Armor NVGs.hpp
 
 #include "urban\config_weapons.hpp"
 #include "arid\config_weapons.hpp"
 #include "arctic\config_weapons.hpp"
 #include "tropic\config_weapons.hpp"
 #include "woodland\config_weapons.hpp"
+#include "specialist_helmets\config_weapons.hpp"
