+++ src/Addons/34thPRC_ArmourStandard/data/marines2/specialist_helmets/config_weapons.hpp
@@ -1,0 +1,1 @@
+#include "EOD\config_weapons.hpp"
